Split previous class into seperate tfidf and nb classes. Added some defaults and automation to ensure smoother running by novices
